[#14] OpenVidu Video 채팅방 내부 동작 확인
css 변경 필요

Resolves : #14
